Deleted - TransactionEntity, TransactionSearchStrategy
With

@mrlumbu
