--- src/main/java/uk/gov/pay/connector/model/domain/transaction/TransactionOperation.java
@@ -1,6 +1,0 @@
-package uk.gov.pay.connector.model.domain.transaction;
-
-public enum TransactionOperation {
-    CHARGE,
-    REFUND
-}
