+++ src/main/java/uk/gov/pay/connector/service/ChargeRefundService.java
@@ -13,10 +13,8 @@ import uk.gov.pay.connector.model.RefundRequest;
 import uk.gov.pay.connector.model.api.ExternalChargeRefundAvailability;
 import uk.gov.pay.connector.model.domain.ChargeEntity;
 import uk.gov.pay.connector.model.domain.GatewayAccountEntity;
-import uk.gov.pay.connector.model.domain.PaymentRequestEntity;
 import uk.gov.pay.connector.model.domain.RefundEntity;
 import uk.gov.pay.connector.model.domain.RefundStatus;
-import uk.gov.pay.connector.model.domain.transaction.RefundTransactionEntity;
 import uk.gov.pay.connector.model.gateway.GatewayResponse;
 import uk.gov.pay.connector.service.transaction.NonTransactionalOperation;
 import uk.gov.pay.connector.service.transaction.PreTransactionalOperation;
@@ -59,17 +57,15 @@ public class ChargeRefundService {
     private final RefundDao refundDao;
     private final PaymentProviders providers;
     private final Provider<TransactionFlow> transactionFlowProvider;
-    private final RefundStatusUpdater refundStatusUpdater;
 
     @Inject
     public ChargeRefundService(ChargeDao chargeDao, RefundDao refundDao, PaymentProviders providers,
-                               Provider<TransactionFlow> transactionFlowProvider, 
-                               RefundStatusUpdater refundStatusUpdater) {
+                               Provider<TransactionFlow> transactionFlowProvider
+                               ) {
         this.chargeDao = chargeDao;
         this.refundDao = refundDao;
         this.providers = providers;
         this.transactionFlowProvider = transactionFlowProvider;
-        this.refundStatusUpdater = refundStatusUpdater;
     }
 
     public Optional<Response> doRefund(Long accountId, String chargeId, RefundRequest refundRequest) {
@@ -90,7 +86,6 @@ public class ChargeRefundService {
                     && refund.hasStatus(RefundStatus.REFUND_SUBMITTED)) {
                 RefundEntity refundEntity = refundDao.findById(refund.getId()).get();
                 refundEntity.setStatus(REFUNDED);
-                refundStatusUpdater.updateRefundTransactionStatus(PaymentGatewayName.SANDBOX, refund.getReference(), REFUNDED);
                 response = new Response(response.getRefundGatewayResponse(), refundEntity);
             }
             return response;
@@ -153,7 +148,6 @@ public class ChargeRefundService {
 
             refundEntity.setStatus(status);
             refundEntity.setReference(reference);
-            refundStatusUpdater.setReferenceAndUpdateTransactionStatus(refundEntity.getExternalId(), reference, status);
 
             return new Response(gatewayResponse, refundEntity);
         };
