+++ src/main/java/uk/gov/pay/connector/service/NotificationService.java
@@ -37,17 +37,15 @@ public class NotificationService {
     private final PaymentProviders paymentProviders;
     private final DnsUtils dnsUtils;
     private final ChargeStatusUpdater chargeStatusUpdater;
-    private final RefundStatusUpdater refundStatusUpdater;
 
     @Inject
-    public NotificationService(ChargeDao chargeDao, ChargeEventDao chargeEventDao, RefundDao refundDao, PaymentProviders paymentProviders, DnsUtils dnsUtils, ChargeStatusUpdater chargeStatusUpdater, RefundStatusUpdater refundStatusUpdater) {
+    public NotificationService(ChargeDao chargeDao, ChargeEventDao chargeEventDao, RefundDao refundDao, PaymentProviders paymentProviders, DnsUtils dnsUtils, ChargeStatusUpdater chargeStatusUpdater) {
         this.chargeDao = chargeDao;
         this.chargeEventDao = chargeEventDao;
         this.refundDao = refundDao;
         this.paymentProviders = paymentProviders;
         this.dnsUtils = dnsUtils;
         this.chargeStatusUpdater = chargeStatusUpdater;
-        this.refundStatusUpdater = refundStatusUpdater;
     }
 
     @Transactional
@@ -240,9 +238,7 @@ public class NotificationService {
             RefundStatus newStatus = notification.getRefundStatus();
 
             refundEntity.setStatus(newStatus);
-            refundStatusUpdater.updateRefundTransactionStatus(
-                    paymentProvider.getPaymentGatewayName(), notification.getReference(), newStatus
-            );
+            
             GatewayAccountEntity gatewayAccount = refundEntity.getChargeEntity().getGatewayAccount();
             logger.info("Notification received for refund. Updating refund - charge_external_id={}, refund_reference={}, transaction_id={}, status={}, "
                             + "status_to={}, account_id={}, provider={}, provider_type={}",
