+++ src/main/java/uk/gov/pay/connector/service/search/TransactionSearchStrategy.java
@@ -1,10 +1,16 @@
 package uk.gov.pay.connector.service.search;
 
+import java.util.List;
+import java.util.Map;
+import javax.ws.rs.core.UriInfo;
+
+import com.google.inject.Inject;
 import uk.gov.pay.connector.dao.CardTypeDao;
 import uk.gov.pay.connector.dao.ChargeSearchParams;
 import uk.gov.pay.connector.dao.TransactionDao;
 import uk.gov.pay.connector.model.ChargeResponse;
-import uk.gov.pay.connector.model.TransactionResponse;
+import uk.gov.pay.connector.model.ChargeResponse.RefundSummary;
+import uk.gov.pay.connector.model.TransactionResponse.TransactionResponseBuilder;
 import uk.gov.pay.connector.model.TransactionType;
 import uk.gov.pay.connector.model.api.ExternalChargeState;
 import uk.gov.pay.connector.model.api.ExternalRefundStatus;
@@ -13,73 +19,77 @@ import uk.gov.pay.connector.model.domain.ChargeStatus;
 import uk.gov.pay.connector.model.domain.PersistedCard;
 import uk.gov.pay.connector.model.domain.RefundStatus;
 import uk.gov.pay.connector.model.domain.Transaction;
-import uk.gov.pay.connector.model.domain.transaction.TransactionOperation;
 import uk.gov.pay.connector.util.DateTimeUtils;
 
-import javax.inject.Inject;
-import javax.ws.rs.core.UriInfo;
-import java.util.List;
-import java.util.Map;
-
 import static javax.ws.rs.HttpMethod.GET;
+import static uk.gov.pay.connector.model.TransactionResponse.aTransactionResponseBuilder;
 
-public class TransactionSearchStrategy extends AbstractSearchStrategy<Transaction> {
+public class TransactionSearchStrategy extends AbstractSearchStrategy<Transaction> implements SearchStrategy {
 
-    private final TransactionDao transactionDao;
+    private TransactionDao transactionDao;
 
     @Inject
     public TransactionSearchStrategy(TransactionDao transactionDao, CardTypeDao cardTypeDao) {
         super(cardTypeDao);
-        this.transactionDao = transactionDao;
+        this.transactionDao=transactionDao;
     }
 
     @Override
     protected long getTotalFor(ChargeSearchParams params) {
-        return transactionDao.getTotal(params);
+        return transactionDao.getTotalFor(params.getGatewayAccountId(), params);
     }
 
     @Override
     protected List<Transaction> findAllBy(ChargeSearchParams params) {
-        return transactionDao.search(params);
+        return transactionDao.findAllBy(params.getGatewayAccountId(), params);
     }
 
     @Override
     protected ChargeResponse buildResponse(UriInfo uriInfo, Transaction transaction, Map<String, String> cardBrandToLabel) {
         ExternalTransactionState externalTransactionState;
-        TransactionType transactionType;
-        if (transaction.getTransactionType().equals(TransactionOperation.CHARGE.name())) {
-            ExternalChargeState externalChargeState = ChargeStatus.valueOf(ChargeStatus.class, transaction.getStatus()).toExternal();
-            externalTransactionState = new ExternalTransactionState(externalChargeState.getStatus(), externalChargeState.isFinished(), externalChargeState.getCode(), externalChargeState.getMessage());
-            transactionType = TransactionType.PAYMENT;
-        } else {
-            ExternalRefundStatus externalRefundStatus = RefundStatus.valueOf(RefundStatus.class, transaction.getStatus()).toExternal();
+        RefundSummary refundSummary = null;
+        if (TransactionType.REFUND.getValue().equals(transaction.getTransactionType())) {
+            ExternalRefundStatus externalRefundStatus = RefundStatus.fromString(transaction.getStatus()).toExternal();
             externalTransactionState = new ExternalTransactionState(externalRefundStatus.getStatus(), externalRefundStatus.isFinished());
-            transactionType = TransactionType.REFUND;
+            refundSummary = buildRefundSummary(transaction);
+        } else {
+            ExternalChargeState externalChargeState = ChargeStatus.fromString(transaction.getStatus()).toExternal();
+            externalTransactionState = new ExternalTransactionState(externalChargeState.getStatusV2(), externalChargeState.isFinished(), externalChargeState.getCode(), externalChargeState.getMessage());
         }
 
         PersistedCard cardDetails = new PersistedCard();
-        cardDetails.setCardBrand(cardBrandToLabel.get(transaction.getCardBrand()));
+        cardDetails.setCardBrand(transaction.getCardBrandLabel());
         cardDetails.setCardHolderName(transaction.getCardHolderName());
         cardDetails.setExpiryDate(transaction.getExpiryDate());
         cardDetails.setLastDigitsCardNumber(transaction.getLastDigitsCardNumber());
 
-        return TransactionResponse.aTransactionResponseBuilder()
-                .withChargeId(transaction.getExternalId())
+        TransactionResponseBuilder transactionResponseBuilder = aTransactionResponseBuilder()
+                .withTransactionType(transaction.getTransactionType())
                 .withAmount(transaction.getAmount())
-                .withReference(transaction.getReference())
-                .withDescription(transaction.getDescription())
                 .withState(externalTransactionState)
-                .withGatewayTransactionId(transaction.getGatewayTransactionId())
-                .withCreatedDate(DateTimeUtils.toUTCDateTimeString(transaction.getCreatedDate()))
-                .withEmail(transaction.getEmail())
                 .withCardDetails(cardDetails)
-                .withTransactionType(transactionType.getValue())
+                .withChargeId(transaction.getExternalId())
+                .withCreatedDate(DateTimeUtils.toUTCDateTimeString(transaction.getCreatedDate()))
+                .withDescription(transaction.getDescription())
+                .withReference(transaction.getReference())
+                .withEmail(transaction.getEmail())
+                .withGatewayTransactionId(transaction.getGatewayTransactionId())
                 .withLink("self", GET, uriInfo.getBaseUriBuilder()
                         .path("/v1/api/accounts/{accountId}/charges/{chargeId}")
                         .build(transaction.getGatewayAccountId(), transaction.getExternalId()))
                 .withLink("refunds", GET, uriInfo.getBaseUriBuilder()
                         .path("/v1/api/accounts/{accountId}/charges/{chargeId}/refunds")
-                        .build(transaction.getGatewayAccountId(), transaction.getExternalId()))
-                .build();
+                        .build(transaction.getGatewayAccountId(), transaction.getExternalId()));
+        if (refundSummary != null) {
+            transactionResponseBuilder.withRefunds(refundSummary);
+        }
+        return transactionResponseBuilder.build();
+    }
+
+    private ChargeResponse.RefundSummary buildRefundSummary(Transaction transaction) {
+        ChargeResponse.RefundSummary refund = new ChargeResponse.RefundSummary();
+        refund.setStatus(transaction.getStatus());
+        refund.setUserExternalId(transaction.getUserExternalId());
+        return refund;
     }
 }
