+++ src/test/java/uk/gov/pay/connector/it/dao/TransactionDaoITest.java
@@ -3,576 +3,1191 @@ package uk.gov.pay.connector.it.dao;
 import org.junit.Before;
 import org.junit.Test;
 import uk.gov.pay.connector.dao.ChargeSearchParams;
-import uk.gov.pay.connector.dao.PaymentRequestDao;
 import uk.gov.pay.connector.dao.TransactionDao;
 import uk.gov.pay.connector.model.ServicePaymentReference;
-import uk.gov.pay.connector.model.TransactionType;
 import uk.gov.pay.connector.model.domain.ChargeStatus;
-import uk.gov.pay.connector.model.domain.GatewayAccountEntity;
-import uk.gov.pay.connector.model.domain.PaymentRequestEntity;
 import uk.gov.pay.connector.model.domain.RefundStatus;
 import uk.gov.pay.connector.model.domain.Transaction;
-import uk.gov.pay.connector.model.domain.transaction.TransactionOperation;
+import uk.gov.pay.connector.util.DateTimeUtils;
 
 import java.time.ZonedDateTime;
-import java.util.HashMap;
+import java.time.temporal.ChronoUnit;
+import java.util.Arrays;
 import java.util.List;
 
-import static java.util.Arrays.asList;
+import static java.time.ZonedDateTime.now;
 import static java.util.Collections.singletonList;
+import static org.exparity.hamcrest.date.ZonedDateTimeMatchers.within;
 import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.nullValue;
 import static org.hamcrest.core.Is.is;
+import static uk.gov.pay.connector.it.dao.DatabaseFixtures.withDatabaseTestHelper;
+import static uk.gov.pay.connector.model.TransactionType.PAYMENT;
 import static uk.gov.pay.connector.model.TransactionType.REFUND;
-import static uk.gov.pay.connector.model.domain.GatewayAccountEntity.Type.TEST;
-import static uk.gov.pay.connector.model.domain.PaymentRequestEntityFixture.aValidPaymentRequestEntity;
-import static uk.gov.pay.connector.model.domain.PaymentRequestEntityFixture.aValidPaymentRequestEntityWithRefund;
-import static uk.gov.pay.connector.model.domain.transaction.ChargeTransactionEntityBuilder.aChargeTransactionEntity;
+import static uk.gov.pay.connector.model.api.ExternalChargeState.EXTERNAL_CREATED;
+import static uk.gov.pay.connector.model.api.ExternalChargeState.EXTERNAL_STARTED;
+import static uk.gov.pay.connector.model.api.ExternalRefundStatus.EXTERNAL_SUBMITTED;
+import static uk.gov.pay.connector.model.api.ExternalRefundStatus.EXTERNAL_SUCCESS;
+import static uk.gov.pay.connector.model.domain.ChargeStatus.AUTHORISATION_READY;
+import static uk.gov.pay.connector.model.domain.ChargeStatus.ENTERING_CARD_DETAILS;
+import static uk.gov.pay.connector.model.domain.RefundStatus.CREATED;
+import static uk.gov.pay.connector.model.domain.RefundStatus.REFUNDED;
 
 public class TransactionDaoITest extends DaoITestBase {
-    private DatabaseFixtures.TestAccount defaultTestAccount;
-    private GatewayAccountEntity gatewayAccount;
-    private PaymentRequestDao paymentRequestDao;
+
+    private static final String FROM_DATE = "2016-01-01T01:00:00Z";
+    private static final String TO_DATE = "2026-01-08T01:00:00Z";
+    private static final String DEFAULT_TEST_CHARGE_DESCRIPTION = "Charge description";
+    private static final String DEFAULT_TEST_CHARGE_REFERENCE = "Council tax thing";
+    private static final String REFUND_USER_EXTERNAL_ID = "user";
     private TransactionDao transactionDao;
+    private DatabaseFixtures.TestAccount defaultTestAccount;
 
     @Before
-    public void setUp() {
-        paymentRequestDao = env.getInstance(PaymentRequestDao.class);
-        transactionDao = env.getInstance(TransactionDao.class);
+    public void setupAccountWithOneChargeAndRefundForTheCharge() {
+        transactionDao= env.getInstance(TransactionDao.class);
 
-        insertTestAccount();
-
-        gatewayAccount = new GatewayAccountEntity(
-                defaultTestAccount.getPaymentProvider(), new HashMap<>(), TEST);
-        gatewayAccount.setId(defaultTestAccount.getAccountId());
+        defaultTestAccount =
+                withDatabaseTestHelper(databaseTestHelper)
+                        .aTestAccount()
+                        .insert();
     }
 
     @Test
-    public void shouldReturnTransactions_byGatewayAccountId() {
-        paymentRequestDao.persist(aValidPaymentRequestEntity()
-                .withGatewayAccountEntity(gatewayAccount).build());
-        paymentRequestDao.persist(aValidPaymentRequestEntity()
-                .withGatewayAccountEntity(gatewayAccount).build());
+    public void searchChargesByGatewayAccountIdOnly()  {
 
-        ChargeSearchParams searchParams = createSearchParams();
-        Long expectedId = searchParams.getGatewayAccountId();
+        // given
+        DatabaseFixtures.TestCharge testCharge = insertNewChargeWithId(1L, now());
+        DatabaseFixtures.TestCardDetails testCardDetails = updateCardDetailsForCharge(testCharge);
+        DatabaseFixtures.TestRefund testRefund = insertNewRefundForCharge(testCharge, REFUND_USER_EXTERNAL_ID, 2L, now().plusSeconds(1));
 
-        final List<Transaction> entityList = transactionDao.search(searchParams);
-        assertThat(entityList.size(), is(2));
-        assertThat(entityList.get(0).getGatewayAccountId(), is(expectedId));
-        assertThat(entityList.get(1).getGatewayAccountId(), is(expectedId));
+        ChargeSearchParams params = new ChargeSearchParams();
+
+        // when
+        List<Transaction> transactions = transactionDao.findAllBy(defaultTestAccount.getAccountId(), params);
+
+        // then
+        assertThat(transactions.size(), is(2));
+
+        Transaction transactionRefund = transactions.get(0);
+        assertThat(transactionRefund.getTransactionType(), is("refund"));
+        assertThat(transactionRefund.getChargeId(), is(testCharge.getChargeId()));
+        assertThat(transactionRefund.getAmount(), is(testRefund.getAmount()));
+        assertThat(transactionRefund.getReference(), is(testCharge.getReference()));
+        assertThat(transactionRefund.getDescription(), is(testCharge.getDescription()));
+        assertThat(transactionRefund.getStatus(), is(testRefund.getStatus().toString()));
+        assertThat(transactionRefund.getCardBrand(), is(testCardDetails.getCardBrand()));
+        assertThat(transactionRefund.getUserExternalId(), is(REFUND_USER_EXTERNAL_ID));
+        assertThat(transactionRefund.getCardBrandLabel(), is("Visa")); // read from card types table which is populated by the card_types.csv seed data
+        assertDateMatch(transactionRefund.getCreatedDate().toString());
+
+        Transaction transactionCharge = transactions.get(1);
+        assertThat(transactionCharge.getTransactionType(), is("charge"));
+        assertThat(transactionCharge.getChargeId(), is(testCharge.getChargeId()));
+        assertThat(transactionCharge.getAmount(), is(testCharge.getAmount()));
+        assertThat(transactionCharge.getReference(), is(testCharge.getReference()));
+        assertThat(transactionCharge.getDescription(), is(testCharge.getDescription()));
+        assertThat(transactionCharge.getStatus(), is(testCharge.getChargeStatus().toString()));
+        assertThat(transactionCharge.getCardBrand(), is(testCardDetails.getCardBrand()));
+        assertThat(transactionCharge.getUserExternalId(), is(nullValue()));
+        assertDateMatch(transactionCharge.getCreatedDate().toString());
     }
 
     @Test
-    public void shouldCountTransactions_byGatewayAccountId() {
-        paymentRequestDao.persist(aValidPaymentRequestEntity()
-                .withGatewayAccountEntity(gatewayAccount).build());
-        paymentRequestDao.persist(aValidPaymentRequestEntity()
-                .withGatewayAccountEntity(gatewayAccount).build());
+    public void searchChargesByFullEmailMatch()  {
 
-        ChargeSearchParams searchParams = new ChargeSearchParams();
-        final Long expectedId = gatewayAccount.getId();
-        searchParams.withGatewayAccountId(expectedId);
+        // given
+        DatabaseFixtures.TestCharge testCharge = insertNewChargeWithId(1L, now().plusHours(1));
+        DatabaseFixtures.TestRefund testRefund = insertNewRefundForCharge(testCharge, REFUND_USER_EXTERNAL_ID,2L, now().plusHours(2));
+        ChargeSearchParams params = new ChargeSearchParams()
+                .withEmailLike(testCharge.getEmail());
 
-        final Long total = transactionDao.getTotal(searchParams);
-        assertThat(total, is(2L));
+        // when
+        List<Transaction> transactions = transactionDao.findAllBy(defaultTestAccount.getAccountId(), params);
+
+        // then
+        assertThat(transactions.size(), is(2));
+
+        assertThat(transactions.get(0).getAmount(), is(testRefund.getAmount()));
+        assertThat(transactions.get(0).getTransactionType(), is("refund"));
+        assertThat(transactions.get(0).getUserExternalId(), is(REFUND_USER_EXTERNAL_ID));
+        assertThat(transactions.get(1).getAmount(), is(testCharge.getAmount()));
+        assertThat(transactions.get(1).getTransactionType(), is("charge"));
+        assertThat(transactions.get(1).getUserExternalId(), is(nullValue()));
     }
 
     @Test
-    public void shouldCountTransactions_GetMoreThanDisplaySize() {
-        paymentRequestDao.persist(aValidPaymentRequestEntity()
-                .withGatewayAccountEntity(gatewayAccount).build());
-        paymentRequestDao.persist(aValidPaymentRequestEntity()
-                .withGatewayAccountEntity(gatewayAccount).build());
+    public void searchChargesByPartialEmailMatch()  {
 
-        ChargeSearchParams searchParams = new ChargeSearchParams();
-        final Long expectedId = gatewayAccount.getId();
-        searchParams.withGatewayAccountId(expectedId);
-        searchParams.withDisplaySize(1L);
+        // given
+        DatabaseFixtures.TestCharge testCharge = insertNewChargeWithId(1L, now().plusHours(1));
+        DatabaseFixtures.TestRefund testRefund = insertNewRefundForCharge(testCharge, REFUND_USER_EXTERNAL_ID,2L, now().plusHours(2));
+        ChargeSearchParams params = new ChargeSearchParams()
+                .withEmailLike("alice");
 
-        final Long total = transactionDao.getTotal(searchParams);
-        assertThat(total, is(2L));
+        // when
+        List<Transaction> transactions = transactionDao.findAllBy(defaultTestAccount.getAccountId(), params);
+
+        // then
+        assertThat(transactions.size(), is(2));
+
+        assertThat(transactions.get(0).getAmount(), is(testRefund.getAmount()));
+        assertThat(transactions.get(0).getTransactionType(), is("refund"));
+        assertThat(transactions.get(0).getUserExternalId(), is(REFUND_USER_EXTERNAL_ID));
+        assertThat(transactions.get(1).getAmount(), is(testCharge.getAmount()));
+        assertThat(transactions.get(1).getTransactionType(), is("charge"));
+        assertThat(transactions.get(1).getUserExternalId(), is(nullValue()));
     }
 
     @Test
-    public void shouldCountTransactions_IncludesBeforeCurrentPage() {
-        paymentRequestDao.persist(aValidPaymentRequestEntity()
-                .withGatewayAccountEntity(gatewayAccount).build());
-        paymentRequestDao.persist(aValidPaymentRequestEntity()
-                .withGatewayAccountEntity(gatewayAccount).build());
+    public void searchChargesByCardBrandOnly()  {
 
-        ChargeSearchParams searchParams = new ChargeSearchParams();
-        final Long expectedId = gatewayAccount.getId();
-        searchParams.withGatewayAccountId(expectedId);
-        searchParams.withDisplaySize(1L);
-        searchParams.withPage(2L);
+        // given
+        DatabaseFixtures.TestCharge testCharge = insertNewChargeWithId(1L, now().plusHours(1));
+        DatabaseFixtures.TestCardDetails testCardDetails = updateCardDetailsForCharge(testCharge);
+        DatabaseFixtures.TestRefund testRefund = insertNewRefundForCharge(testCharge, REFUND_USER_EXTERNAL_ID,2L, now().plusHours(2));
+        ChargeSearchParams params = new ChargeSearchParams()
+                .withCardBrand(testCardDetails.getCardBrand());
 
-        final Long total = transactionDao.getTotal(searchParams);
-        assertThat(total, is(2L));
+        // when
+        List<Transaction> transactions = transactionDao.findAllBy(defaultTestAccount.getAccountId(), params);
+
+        // then
+        assertThat(transactions.size(), is(2));
+
+        assertThat(transactions.get(0).getAmount(), is(testRefund.getAmount()));
+        assertThat(transactions.get(0).getTransactionType(), is("refund"));
+        assertThat(transactions.get(0).getUserExternalId(), is(REFUND_USER_EXTERNAL_ID));
+        assertThat(transactions.get(1).getAmount(), is(testCharge.getAmount()));
+        assertThat(transactions.get(1).getTransactionType(), is("charge"));
+        assertThat(transactions.get(1).getUserExternalId(), is(nullValue()));
     }
 
     @Test
-    public void shouldReturnTransactions_byGatewayAccountId_whenMultipleGatewayAccountsExist() {
-        final PaymentRequestEntity paymentRequestEntity = aValidPaymentRequestEntity()
-                .withGatewayAccountEntity(gatewayAccount).build();
-        paymentRequestDao.persist(paymentRequestEntity);
+    public void searchChargesByMultipleCardBrandOnly()  {
 
-        DatabaseFixtures.TestAccount defaultTestAccount2 = DatabaseFixtures
-                .withDatabaseTestHelper(databaseTestHelper).aTestAccount().insert();
-        GatewayAccountEntity gatewayAccount2 = new GatewayAccountEntity(
-                defaultTestAccount2.getPaymentProvider(), new HashMap<>(), TEST);
-        gatewayAccount2.setId(defaultTestAccount2.getAccountId());
+        // given
+        String visa = "visa";
+        String masterCard = "master-card";
+        DatabaseFixtures.TestCharge testCharge1 = insertNewChargeWithId(1L, now().plusHours(1));
+        updateCardDetailsForCharge(testCharge1, visa);
+        DatabaseFixtures.TestRefund testRefund1 = insertNewRefundForCharge(testCharge1, REFUND_USER_EXTERNAL_ID,2L, now().plusHours(2));
 
-        paymentRequestDao.persist(aValidPaymentRequestEntity()
-                .withGatewayAccountEntity(gatewayAccount2).build());
+        DatabaseFixtures.TestCharge testCharge2 = insertNewChargeWithId(3L, now().plusHours(3));
+        updateCardDetailsForCharge(testCharge2, visa);
+        DatabaseFixtures.TestRefund testRefund2 = insertNewRefundForCharge(testCharge2, REFUND_USER_EXTERNAL_ID,4L, now().plusHours(4));
 
-        ChargeSearchParams searchParams = createSearchParams();
+        ChargeSearchParams params = new ChargeSearchParams()
+                .withCardBrands(Arrays.asList(visa, masterCard));
 
-        assertTransactionByExternalId(paymentRequestEntity.getExternalId(), transactionDao.search(searchParams));
+        // when
+        List<Transaction> transactions = transactionDao.findAllBy(defaultTestAccount.getAccountId(), params);
+
+        // then
+        assertThat(transactions.size(), is(4));
+
+        assertThat(transactions.get(0).getAmount(), is(testRefund2.getAmount()));
+        assertThat(transactions.get(0).getTransactionType(), is("refund"));
+        assertThat(transactions.get(0).getUserExternalId(), is(REFUND_USER_EXTERNAL_ID));
+        assertThat(transactions.get(1).getAmount(), is(testCharge2.getAmount()));
+        assertThat(transactions.get(1).getTransactionType(), is("charge"));
+        assertThat(transactions.get(1).getUserExternalId(), is(nullValue()));
+        assertThat(transactions.get(2).getAmount(), is(testRefund1.getAmount()));
+        assertThat(transactions.get(2).getTransactionType(), is("refund"));
+        assertThat(transactions.get(2).getUserExternalId(), is(REFUND_USER_EXTERNAL_ID));
+        assertThat(transactions.get(3).getAmount(), is(testCharge1.getAmount()));
+        assertThat(transactions.get(3).getTransactionType(), is("charge"));
+        assertThat(transactions.get(3).getUserExternalId(), is(nullValue()));
     }
 
     @Test
-    public void shouldReturnTransactions_byEmail_whenMultipleTransactionsWithDifferentEmails() {
-        PaymentRequestEntity paymentRequestEntity1 = aValidPaymentRequestEntity()
-                .withGatewayAccountEntity(gatewayAccount)
-                .build();
+    public void searchChargesWithDefaultSizeAndPage_shouldGetChargesInCreationDateOrder()  {
 
-        final String expectedEmail = "abc@example.com";
-        paymentRequestEntity1.getChargeTransaction().setEmail(expectedEmail);
-        paymentRequestDao.persist(paymentRequestEntity1);
+        // given
+        DatabaseFixtures.TestCharge charge1 = insertNewChargeWithId(1L, now().plusHours(1));
+        insertNewRefundForCharge(charge1, REFUND_USER_EXTERNAL_ID,2L, now().plusHours(2));
+        DatabaseFixtures.TestCharge charge2 = insertNewChargeWithId(3L, now().plusHours(3));
+        insertNewRefundForCharge(charge2, REFUND_USER_EXTERNAL_ID,4L, now().plusHours(4));
+        insertNewRefundForCharge(charge2, REFUND_USER_EXTERNAL_ID,5L, now().plusHours(5));
+        DatabaseFixtures.TestCharge charge3 = insertNewChargeWithId(6L, now().plusHours(6));
 
-        final PaymentRequestEntity paymentRequestEntity2 = aValidPaymentRequestEntity()
-                .withGatewayAccountEntity(gatewayAccount).build();
-        paymentRequestEntity2.getChargeTransaction().setEmail("zzz@example.com");
-        paymentRequestDao.persist(paymentRequestEntity2);
+        ChargeSearchParams params = new ChargeSearchParams();
 
-        ChargeSearchParams searchParams = createSearchParams();
-        searchParams.withEmailLike(expectedEmail);
+        // when
+        List<Transaction> transactions = transactionDao.findAllBy(defaultTestAccount.getAccountId(), params);
 
-        assertTransactionByExternalId(paymentRequestEntity1.getExternalId(), transactionDao.search(searchParams));
+        // then
+        assertThat(transactions.size(), is(6));
+        assertThat(transactions.get(0).getExternalId(), is(charge3.getExternalChargeId()));
+        assertThat(transactions.get(0).getTransactionType(), is("charge"));
+        assertThat(transactions.get(0).getAmount(), is(6L));
+        assertThat(transactions.get(0).getUserExternalId(), is(nullValue()));
+
+        assertThat(transactions.get(1).getExternalId(), is(charge2.getExternalChargeId()));
+        assertThat(transactions.get(1).getTransactionType(), is("refund"));
+        assertThat(transactions.get(1).getAmount(), is(5L));
+        assertThat(transactions.get(1).getUserExternalId(), is(REFUND_USER_EXTERNAL_ID));
+
+        assertThat(transactions.get(2).getExternalId(), is(charge2.getExternalChargeId()));
+        assertThat(transactions.get(2).getTransactionType(), is("refund"));
+        assertThat(transactions.get(2).getAmount(), is(4L));
+        assertThat(transactions.get(2).getUserExternalId(), is(REFUND_USER_EXTERNAL_ID));
+
+        assertThat(transactions.get(3).getExternalId(), is(charge2.getExternalChargeId()));
+        assertThat(transactions.get(3).getTransactionType(), is("charge"));
+        assertThat(transactions.get(3).getAmount(), is(3L));
+        assertThat(transactions.get(3).getUserExternalId(), is(nullValue()));
+        
+        assertThat(transactions.get(4).getExternalId(), is(charge1.getExternalChargeId()));
+        assertThat(transactions.get(4).getTransactionType(), is("refund"));
+        assertThat(transactions.get(4).getAmount(), is(2L));
+        assertThat(transactions.get(4).getUserExternalId(), is(REFUND_USER_EXTERNAL_ID));
+        
+        assertThat(transactions.get(5).getExternalId(), is(charge1.getExternalChargeId()));
+        assertThat(transactions.get(5).getTransactionType(), is("charge"));
+        assertThat(transactions.get(5).getAmount(), is(1L));
+        assertThat(transactions.get(5).getUserExternalId(), is(nullValue()));
     }
 
     @Test
-    public void shouldReturnTransactions_byEmailWithDifferentCasing() {
-        PaymentRequestEntity paymentRequestEntity = aValidPaymentRequestEntity()
-                .withGatewayAccountEntity(gatewayAccount)
-                .build();
+    public void searchChargesWithSizeAndPageSetShouldGetTransactionsInCreationDateOrder()  {
 
-        final String expectedEmail = "Abc@example.com";
-        paymentRequestEntity.getChargeTransaction().setEmail(expectedEmail);
-        paymentRequestDao.persist(paymentRequestEntity);
+        // given
+        DatabaseFixtures.TestCharge charge1 = insertNewChargeWithId(1L, now().plusHours(1));
+        insertNewRefundForCharge(charge1, REFUND_USER_EXTERNAL_ID,2L, now().plusHours(2));
+        DatabaseFixtures.TestCharge charge2 = insertNewChargeWithId(3L, now().plusHours(3));
+        insertNewRefundForCharge(charge2, REFUND_USER_EXTERNAL_ID,4L, now().plusHours(4));
+        insertNewRefundForCharge(charge2, REFUND_USER_EXTERNAL_ID,5L, now().plusHours(5));
+        DatabaseFixtures.TestCharge charge3 = insertNewChargeWithId(6L, now().plusHours(6));
+        DatabaseFixtures.TestCharge charge4 = insertNewChargeWithId(6L, now().plusHours(7));
+        insertNewRefundForCharge(charge3, REFUND_USER_EXTERNAL_ID,5L, now().plusHours(8));
 
-        ChargeSearchParams searchParams = createSearchParams();
-        searchParams.withEmailLike(expectedEmail.toLowerCase());
+        // when
+        ChargeSearchParams params = new ChargeSearchParams()
+                .withPage(1L)
+                .withDisplaySize(3L);
 
-        assertTransactionByExternalId(paymentRequestEntity.getExternalId(), transactionDao.search(searchParams));
+        List<Transaction> transactions = transactionDao.findAllBy(defaultTestAccount.getAccountId(), params);
+        // then
+        assertThat(transactions.size(), is(3));
+
+        assertThat(transactions.get(0).getExternalId(), is(charge3.getExternalChargeId()));
+        assertThat(transactions.get(0).getTransactionType(), is("refund"));
+        assertThat(transactions.get(0).getUserExternalId(), is(REFUND_USER_EXTERNAL_ID));
+
+        assertThat(transactions.get(1).getExternalId(), is(charge4.getExternalChargeId()));
+        assertThat(transactions.get(1).getTransactionType(), is("charge"));
+        assertThat(transactions.get(1).getUserExternalId(), is(nullValue()));
+
+        assertThat(transactions.get(2).getExternalId(), is(charge3.getExternalChargeId()));
+        assertThat(transactions.get(2).getTransactionType(), is("charge"));
+        assertThat(transactions.get(2).getUserExternalId(), is(nullValue()));
+
+        // when
+        params = new ChargeSearchParams()
+                .withGatewayAccountId(defaultTestAccount.getAccountId())
+                .withPage(2L)
+                .withDisplaySize(3L);
+
+        transactions = transactionDao.findAllBy(defaultTestAccount.getAccountId(), params);
+
+        // then
+        assertThat(transactions.size(), is(3));
+
+        assertThat(transactions.get(0).getExternalId(), is(charge2.getExternalChargeId()));
+        assertThat(transactions.get(0).getTransactionType(), is("refund"));
+        assertThat(transactions.get(0).getAmount(), is(5L));
+        assertThat(transactions.get(0).getUserExternalId(), is(REFUND_USER_EXTERNAL_ID));
+
+        assertThat(transactions.get(1).getExternalId(), is(charge2.getExternalChargeId()));
+        assertThat(transactions.get(1).getTransactionType(), is("refund"));
+        assertThat(transactions.get(1).getAmount(), is(4L));
+        assertThat(transactions.get(1).getUserExternalId(), is(REFUND_USER_EXTERNAL_ID));
+
+        assertThat(transactions.get(2).getExternalId(), is(charge2.getExternalChargeId()));
+        assertThat(transactions.get(2).getTransactionType(), is("charge"));
+        assertThat(transactions.get(2).getUserExternalId(), is(nullValue()));
+
+
+        // when
+        params = new ChargeSearchParams()
+                .withGatewayAccountId(defaultTestAccount.getAccountId())
+                .withPage(3L)
+                .withDisplaySize(3L);
+
+        transactions = transactionDao.findAllBy(defaultTestAccount.getAccountId(), params);
+
+        // then
+        assertThat(transactions.size(), is(2));
+
+        assertThat(transactions.get(0).getExternalId(), is(charge1.getExternalChargeId()));
+        assertThat(transactions.get(0).getTransactionType(), is("refund"));
+        assertThat(transactions.get(0).getUserExternalId(), is(REFUND_USER_EXTERNAL_ID));
+
+        assertThat(transactions.get(1).getExternalId(), is(charge1.getExternalChargeId()));
+        assertThat(transactions.get(1).getTransactionType(), is("charge"));
+        assertThat(transactions.get(1).getUserExternalId(), is(nullValue()));
+
     }
 
     @Test
-    public void shouldReturnTransactions_byEmailWithLowCaseInDb() {
-        PaymentRequestEntity paymentRequestEntity1 = aValidPaymentRequestEntity()
-                .withGatewayAccountEntity(gatewayAccount)
-                .build();
+    public void shouldGetExpectedTotalCount() {
 
-        final String expectedEmail = "abc@example.com";
-        paymentRequestEntity1.getChargeTransaction().setEmail(expectedEmail);
-        paymentRequestDao.persist(paymentRequestEntity1);
+        // given
+        DatabaseFixtures.TestCharge charge1 = insertNewChargeWithId(1L, now().plusHours(1));
+        insertNewRefundForCharge(charge1, REFUND_USER_EXTERNAL_ID,2L, now().plusHours(2));
+        DatabaseFixtures.TestCharge charge2 = insertNewChargeWithId(3L, now().plusHours(3));
+        insertNewRefundForCharge(charge2, REFUND_USER_EXTERNAL_ID,4L, now().plusHours(4));
+        insertNewRefundForCharge(charge2, REFUND_USER_EXTERNAL_ID,5L, now().plusHours(5));
+        DatabaseFixtures.TestCharge charge3 = insertNewChargeWithId(6L, now().plusHours(6));
+        insertNewChargeWithId(6L, now().plusHours(7));
+        insertNewRefundForCharge(charge3, REFUND_USER_EXTERNAL_ID,5L, now().plusHours(8));
 
-        ChargeSearchParams searchParams = createSearchParams();
-        searchParams.withEmailLike(expectedEmail.toUpperCase());
+        ChargeSearchParams params = new ChargeSearchParams()
+                .withPage(1L)
+                .withDisplaySize(2L);
 
-        assertTransactionByExternalId(paymentRequestEntity1.getExternalId(), transactionDao.search(searchParams));
+        // when
+        Long count = transactionDao.getTotalFor(defaultTestAccount.getAccountId(), params);
+
+        // then
+        assertThat("total count for transactions matches", count, is(8L));
+
     }
 
     @Test
-    public void shouldReturnTransactions_byEmailWithPartialMatch() {
-        PaymentRequestEntity paymentRequestEntity = aValidPaymentRequestEntity()
-                .withGatewayAccountEntity(gatewayAccount)
-                .build();
+    public void searchChargesByFullReferenceOnly()  {
 
-        final String expectedEmail = "abc@example.com";
-        paymentRequestEntity.getChargeTransaction().setEmail(expectedEmail);
-        paymentRequestDao.persist(paymentRequestEntity);
+        // given
+        DatabaseFixtures.TestCharge testCharge = insertNewChargeWithId(1L, now().plusHours(1));
+        insertNewRefundForCharge(testCharge, REFUND_USER_EXTERNAL_ID,2L, now().plusHours(2));
+        ChargeSearchParams params = new ChargeSearchParams()
+                .withReferenceLike(testCharge.getReference());
 
-        ChargeSearchParams searchParams = createSearchParams();
-        searchParams.withEmailLike("abc");
+        // when
+        List<Transaction> transactions = transactionDao.findAllBy(defaultTestAccount.getAccountId(), params);
 
-        assertTransactionByExternalId(paymentRequestEntity.getExternalId(), transactionDao.search(searchParams));
+        // then
+        assertThat(transactions.size(), is(2));
+
+        Transaction refund = transactions.get(0);
+        Transaction charge = transactions.get(1);
+
+        assertThat(refund.getExternalId(), is(testCharge.getExternalChargeId()));
+        assertThat(refund.getTransactionType(), is("refund"));
+        assertThat(refund.getUserExternalId(), is(REFUND_USER_EXTERNAL_ID));
+        assertThat(charge.getExternalId(), is(testCharge.getExternalChargeId()));
+        assertThat(charge.getTransactionType(), is("charge"));
+        assertThat(charge.getUserExternalId(), is(nullValue()));
     }
 
     @Test
-    public void shouldReturnRefundTransactions_byEmailWithPartialMatch() {
-        PaymentRequestEntity paymentRequestEntity = aValidPaymentRequestEntityWithRefund()
-                .withGatewayAccountEntity(gatewayAccount)
-                .build();
+    public void searchChargesByPartialReferenceOnly()  {
 
-        final String expectedEmail = "abc@example.com";
-        paymentRequestEntity.getChargeTransaction().setEmail(expectedEmail);
-        paymentRequestDao.persist(paymentRequestEntity);
+        // given
+        DatabaseFixtures.TestCharge testCharge = insertNewChargeWithId(1L, now().plusHours(1));
+        insertNewRefundForCharge(testCharge, REFUND_USER_EXTERNAL_ID,2L, now().plusHours(2));
 
-        ChargeSearchParams searchParams = createSearchParams();
-        searchParams.withEmailLike("abc");
-        searchParams.withTransactionType(REFUND);
+        ServicePaymentReference partialPaymentReference = ServicePaymentReference.of("Council Tax");
 
-        assertTransactionByExternalId(paymentRequestEntity.getExternalId(), transactionDao.search(searchParams));
+        DatabaseFixtures.TestCharge partialReferenceCharge =
+                withDatabaseTestHelper(databaseTestHelper)
+                        .aTestCharge()
+                        .withReference(ServicePaymentReference.of(partialPaymentReference.toString() + " whatever"))
+                        .withTestAccount(defaultTestAccount)
+                        .withCreatedDate(now().plusHours(3))
+                        .insert();
+
+        ChargeSearchParams params = new ChargeSearchParams()
+                .withReferenceLike(partialPaymentReference);
+
+        // when
+        List<Transaction> transactions = transactionDao.findAllBy(defaultTestAccount.getAccountId(), params);
+
+        // then
+        assertThat(transactions.size(), is(3));
+        assertThat(transactions.get(0).getExternalId(), is(partialReferenceCharge.getExternalChargeId()));
+        assertThat(transactions.get(0).getReference(), is(partialReferenceCharge.getReference()));
+        assertThat(transactions.get(0).getTransactionType(), is("charge"));
+        assertThat(transactions.get(0).getUserExternalId(), is(nullValue()));
+
+        assertThat(transactions.get(1).getExternalId(), is(testCharge.getExternalChargeId()));
+        assertThat(transactions.get(1).getTransactionType(), is("refund"));
+        assertThat(transactions.get(1).getUserExternalId(), is(REFUND_USER_EXTERNAL_ID));
+
+        assertThat(transactions.get(2).getExternalId(), is(testCharge.getExternalChargeId()));
+        assertThat(transactions.get(2).getReference(), is(testCharge.getReference()));
+        assertThat(transactions.get(2).getTransactionType(), is("charge"));
+        assertThat(transactions.get(2).getUserExternalId(), is(nullValue()));
+
     }
 
     @Test
-    public void shouldReturnTransactions_byEmailWithUnderscore() {
-        PaymentRequestEntity paymentRequestEntity1 = aValidPaymentRequestEntity()
-                .withGatewayAccountEntity(gatewayAccount)
-                .build();
-
-        final String expectedEmail = "a_c@example.com";
-        paymentRequestEntity1.getChargeTransaction().setEmail(expectedEmail);
-        paymentRequestDao.persist(paymentRequestEntity1);
-
-        final PaymentRequestEntity paymentRequestEntity2 = aValidPaymentRequestEntity()
-                .withGatewayAccountEntity(gatewayAccount).build();
-        paymentRequestEntity2.getChargeTransaction().setEmail("abc@example.com");
-        paymentRequestDao.persist(paymentRequestEntity2);
-
-        ChargeSearchParams searchParams = createSearchParams();
-        searchParams.withEmailLike(expectedEmail);
-
-        assertTransactionByExternalId(paymentRequestEntity1.getExternalId(), transactionDao.search(searchParams));
-    }
-
-    @Test
-    public void shouldReturnTransactions_byEmailWithPercent() {
-        PaymentRequestEntity paymentRequestEntity1 = aValidPaymentRequestEntity()
-                .withGatewayAccountEntity(gatewayAccount)
-                .build();
-
-        final String expectedEmail = "a%c@example.com";
-        paymentRequestEntity1.getChargeTransaction().setEmail(expectedEmail);
-        paymentRequestDao.persist(paymentRequestEntity1);
-
-        final PaymentRequestEntity paymentRequestEntity2 = aValidPaymentRequestEntity()
-                .withGatewayAccountEntity(gatewayAccount).build();
-        paymentRequestEntity2.getChargeTransaction().setEmail("abc@example.com");
-        paymentRequestDao.persist(paymentRequestEntity2);
-
-        ChargeSearchParams searchParams = createSearchParams();
-        searchParams.withEmailLike(expectedEmail);
-
-        assertTransactionByExternalId(paymentRequestEntity1.getExternalId(), transactionDao.search(searchParams));
-    }
-
-    @Test
-    public void shouldReturnTransactions_byReference() {
-        final ServicePaymentReference expectedReference = ServicePaymentReference.of("some random reference");
-        PaymentRequestEntity paymentRequestEntity = aValidPaymentRequestEntity()
-                .withGatewayAccountEntity(gatewayAccount)
-                .withReference(expectedReference)
-                .build();
-
-        paymentRequestDao.persist(paymentRequestEntity);
-
-        paymentRequestDao.persist(aValidPaymentRequestEntity()
-                .withGatewayAccountEntity(gatewayAccount).build());
-
-        ChargeSearchParams searchParams = createSearchParams();
-        searchParams.withReferenceLike(expectedReference);
-
-        assertTransactionByExternalId(paymentRequestEntity.getExternalId(), transactionDao.search(searchParams));
-    }
-
-    @Test
-    public void shouldReturnTransactions_byReferenceWithUnderscore() {
-        final ServicePaymentReference expectedReference = ServicePaymentReference.of("a_cedkdkwd");
-
-        PaymentRequestEntity paymentRequestEntity1 = aValidPaymentRequestEntity()
-                .withGatewayAccountEntity(gatewayAccount)
-                .withReference(expectedReference)
-                .build();
-
-        paymentRequestDao.persist(paymentRequestEntity1);
-
-        final PaymentRequestEntity paymentRequestEntity2 = aValidPaymentRequestEntity()
-                .withGatewayAccountEntity(gatewayAccount).withReference(ServicePaymentReference.of("abcedkdkwd")).build();
-        paymentRequestDao.persist(paymentRequestEntity2);
-
-        ChargeSearchParams searchParams = createSearchParams();
-        searchParams.withReferenceLike(expectedReference);
-
-        assertTransactionByExternalId(paymentRequestEntity1.getExternalId(), transactionDao.search(searchParams));
-    }
-
-    @Test
-    public void shouldReturnTransactions_byReferenceWithPercent() {
-        final ServicePaymentReference expectedReference = ServicePaymentReference.of("a%cedkdkwd");
-
-        PaymentRequestEntity paymentRequestEntity1 = aValidPaymentRequestEntity()
-                .withGatewayAccountEntity(gatewayAccount)
-                .withReference(expectedReference)
-                .build();
-
-        paymentRequestDao.persist(paymentRequestEntity1);
-
-        final PaymentRequestEntity paymentRequestEntity2 = aValidPaymentRequestEntity()
-                .withGatewayAccountEntity(gatewayAccount)
-                .withReference(ServicePaymentReference.of("abcedkdkwd"))
-                .build();
-        paymentRequestDao.persist(paymentRequestEntity2);
-
-        ChargeSearchParams searchParams = createSearchParams();
-        searchParams.withReferenceLike(expectedReference);
-
-        assertTransactionByExternalId(paymentRequestEntity1.getExternalId(), transactionDao.search(searchParams));
-    }
-
-    @Test
-    public void shouldReturnTransactions_byReferenceWithUppercase() {
-        final ServicePaymentReference expectedReference = ServicePaymentReference.of("ABCEdkdkwd");
-
-        PaymentRequestEntity paymentRequestEntity = aValidPaymentRequestEntity()
-                .withGatewayAccountEntity(gatewayAccount)
-                .withReference(expectedReference)
-                .build();
-
-        paymentRequestDao.persist(paymentRequestEntity);
-
-        ChargeSearchParams searchParams = createSearchParams();
-        searchParams.withReferenceLike(expectedReference);
-
-        assertTransactionByExternalId(paymentRequestEntity.getExternalId(), transactionDao.search(searchParams));
-    }
-
-    @Test
-    public void shouldReturnTransactions_byReferenceWithUppercaseInRequest() {
-        final ServicePaymentReference expectedReference = ServicePaymentReference.of("dkdkwd");
-
-        PaymentRequestEntity paymentRequestEntity = aValidPaymentRequestEntity()
-                .withGatewayAccountEntity(gatewayAccount)
-                .withReference(expectedReference)
-                .build();
-
-        paymentRequestDao.persist(paymentRequestEntity);
-
-        ChargeSearchParams searchParams = createSearchParams();
-        searchParams.withReferenceLike(ServicePaymentReference.of(expectedReference.toString().toUpperCase()));
-
-        assertTransactionByExternalId(paymentRequestEntity.getExternalId(), transactionDao.search(searchParams));
-    }
-
-    @Test
-    public void shouldReturnTransactions_byFromDate() {
-        final ZonedDateTime paymentDate = ZonedDateTime.now().minusMinutes(10);
-
-        final PaymentRequestEntity paymentRequestEntity = aValidPaymentRequestEntity()
-                .withGatewayAccountEntity(gatewayAccount)
-                .build();
-        paymentRequestEntity.getChargeTransaction().setCreatedDate(paymentDate);
-        final String expectedExternalId = paymentRequestEntity.getExternalId();
-        paymentRequestDao.persist(paymentRequestEntity);
-
-        final PaymentRequestEntity paymentRequestEntity2 = aValidPaymentRequestEntity()
-                .withGatewayAccountEntity(gatewayAccount)
-                .build();
-        paymentRequestEntity2.getChargeTransaction().setCreatedDate(paymentDate.minusMinutes(5));
-        paymentRequestDao.persist(paymentRequestEntity2);
-
-        ChargeSearchParams searchParams = createSearchParams();
-        searchParams.withFromDate(paymentDate);
-
-        assertTransactionByExternalId(expectedExternalId, transactionDao.search(searchParams));
-    }
-
-    @Test
-    public void shouldReturnTransactions_byToDate() {
-        final ZonedDateTime paymentDate = ZonedDateTime.now().minusMinutes(10);
-
-        final PaymentRequestEntity paymentRequestEntity = aValidPaymentRequestEntity()
-                .withGatewayAccountEntity(gatewayAccount)
-                .build();
-        paymentRequestEntity.getChargeTransaction().setCreatedDate(paymentDate);
-        final String expectedExternalId = paymentRequestEntity.getExternalId();
-        paymentRequestDao.persist(paymentRequestEntity);
-
-        final PaymentRequestEntity paymentRequestEntity2 = aValidPaymentRequestEntity()
-                .withGatewayAccountEntity(gatewayAccount)
-                .build();
-        paymentRequestEntity2.getChargeTransaction().setCreatedDate(paymentDate.plusMinutes(5));
-        paymentRequestDao.persist(paymentRequestEntity2);
-
-        ChargeSearchParams searchParams = createSearchParams();
-        searchParams.withToDate(paymentDate.plusMinutes(1));
-
-        assertTransactionByExternalId(expectedExternalId, transactionDao.search(searchParams));
-    }
-
-    @Test
-    public void shouldReturnRefundTransactions_byTransactionType() {
-        final PaymentRequestEntity paymentRequestEntity = aValidPaymentRequestEntityWithRefund()
-                .withGatewayAccountEntity(gatewayAccount)
-                .build();
-        final String externalId = paymentRequestEntity.getExternalId();
-        paymentRequestDao.persist(paymentRequestEntity);
-
-        ChargeSearchParams searchParams = createSearchParams();
-        searchParams.withTransactionType(REFUND);
-
-        final List<Transaction> searchResult = transactionDao.search(searchParams);
-        assertIsTransactionOperation(searchResult, TransactionOperation.REFUND);
-        assertThat(searchResult.get(0).getExternalId(), is(externalId));
-    }
-
-    @Test
-    public void shouldReturnChargeTransactions_byTransactionType() {
-        final PaymentRequestEntity paymentRequestEntity = aValidPaymentRequestEntityWithRefund()
-                .withGatewayAccountEntity(gatewayAccount)
-                .build();
-        paymentRequestDao.persist(paymentRequestEntity);
-        final long chargeTransactionId = paymentRequestEntity.getChargeTransaction().getId();
-
-        ChargeSearchParams searchParams = createSearchParams();
-        searchParams.withTransactionType(TransactionType.PAYMENT);
-
-        final List<Transaction> searchResult = transactionDao.search(searchParams);
-        assertIsTransactionOperation(searchResult, TransactionOperation.CHARGE);
-        final long actualChargeTransactionId = searchResult.get(0).getChargeId();
-        assertThat(actualChargeTransactionId, is(chargeTransactionId));
-    }
-
-    private void assertIsTransactionOperation(List<Transaction> searchResult, TransactionOperation transactionOperation) {
-        assertThat(searchResult.size(), is(1));
-        assertThat(searchResult.get(0).getTransactionType(), is(transactionOperation.name()));
-    }
-
-    @Test
-    public void shouldReturnTransactions_byStatus() {
-        final ChargeStatus chargeStatus = ChargeStatus.CAPTURED;
-        final PaymentRequestEntity paymentRequestEntity =
-                persistPaymentRequestForStatus(chargeStatus, RefundStatus.REFUNDED);
-
-        final Long chargeTransactionId = paymentRequestEntity.getChargeTransaction().getId();
-
-        ChargeSearchParams searchParams = createSearchParams();
-        searchParams.withExternalState(chargeStatus.toExternal().getStatus());
-
-        assertTransactionById(chargeTransactionId, transactionDao.search(searchParams));
-    }
-
-    @Test
-    public void shouldReturnTransactions_byChargeStatus() {
-        final ChargeStatus chargeStatus = ChargeStatus.CAPTURED;
-        final PaymentRequestEntity paymentRequestEntity =
-                persistPaymentRequestForStatus(chargeStatus, RefundStatus.REFUNDED);
-
-        final Long chargeTransactionId = paymentRequestEntity.getChargeTransaction().getId();
-
-        ChargeSearchParams searchParams = createSearchParams();
-        searchParams.addExternalChargeStates(singletonList(chargeStatus.toExternal().getStatus()));
-
-        assertTransactionById(chargeTransactionId, transactionDao.search(searchParams));
-    }
-
-    @Test
-    public void shouldReturnTransactions_byRefundStatus() {
-        final RefundStatus refundStatus = RefundStatus.REFUNDED;
-        final PaymentRequestEntity paymentRequestEntity =
-                persistPaymentRequestForStatus(ChargeStatus.CAPTURED, refundStatus);
-
-        final Long refundTransactionId = paymentRequestEntity.getRefundTransactions().get(0).getId();
-
-        ChargeSearchParams searchParams = createSearchParams();
-        searchParams.addExternalRefundStates(singletonList(refundStatus.toExternal().getStatus()));
-
-        assertTransactionById(refundTransactionId, transactionDao.search(searchParams));
-    }
-
-    @Test
-    public void shouldReturnTransactions_byMultipleStatues() {
-        final ChargeStatus chargeStatus = ChargeStatus.CAPTURED;
-        final RefundStatus refundStatus = RefundStatus.REFUNDED;
-        final PaymentRequestEntity paymentRequestEntity =
-                persistPaymentRequestForStatus(chargeStatus, refundStatus);
-
-        final Long chargeTransactionId = paymentRequestEntity.getChargeTransaction().getId();
-        final Long refundTransactionId = paymentRequestEntity.getRefundTransactions().get(0).getId();
-
-        ChargeSearchParams searchParams = createSearchParams();
-        searchParams.addExternalChargeStates(singletonList(
-                chargeStatus.toExternal().getStatus()
-        ));
-        searchParams.addExternalRefundStates(singletonList(
-                refundStatus.toExternal().getStatus())
-        );
-
-        final List<Transaction> searchResult = transactionDao.search(searchParams);
-        assertThat(searchResult.size(), is(2));
-        assertThat(searchResult.get(0).getChargeId(), is(refundTransactionId));
-        assertThat(searchResult.get(1).getChargeId(), is(chargeTransactionId));
-    }
-
-    @Test
-    public void shouldReturnTransactions_byDisplaySize() {
-        final PaymentRequestEntity paymentRequestEntity = aValidPaymentRequestEntityWithRefund()
-                .withGatewayAccountEntity(gatewayAccount).build();
-        paymentRequestDao.persist(paymentRequestEntity);
-        paymentRequestDao.persist(aValidPaymentRequestEntityWithRefund()
-                .withGatewayAccountEntity(gatewayAccount).build());
-        paymentRequestDao.persist(aValidPaymentRequestEntityWithRefund()
-                .withGatewayAccountEntity(gatewayAccount).build());
-
-        ChargeSearchParams searchParams = createSearchParams();
-        searchParams.withPage(2L);
-        searchParams.withDisplaySize(4L);
-
-        final List<Transaction> searchResult = transactionDao.search(searchParams);
-        assertThat(searchResult.size(), is(2));
-        final Long refundId = paymentRequestEntity.getRefundTransactions().get(0).getId();
-        assertThat(searchResult.get(0).getChargeId(), is(refundId));
-        final Long chargeId = paymentRequestEntity.getChargeTransaction().getId();
-        assertThat(searchResult.get(1).getChargeId(), is(chargeId));
-    }
-    
-
-    private ChargeSearchParams createSearchParams() {
-        ChargeSearchParams searchParams = new ChargeSearchParams();
-        searchParams.withGatewayAccountId(gatewayAccount.getId());
-        searchParams.withPage(1L);
-        searchParams.withDisplaySize(500L);
-        return searchParams;
-    }
-
-    private PaymentRequestEntity persistPaymentRequestForStatus(ChargeStatus chargeStatus, RefundStatus refundStatus) {
-        PaymentRequestEntity paymentRequestEntity = aValidPaymentRequestEntityWithRefund()
-                .withGatewayAccountEntity(gatewayAccount)
-                .build();
-
-        paymentRequestEntity.getChargeTransaction().updateStatus(chargeStatus);
-        paymentRequestEntity.getRefundTransactions().get(0).updateStatus(refundStatus);
-        paymentRequestDao.persist(paymentRequestEntity);
-
-        return paymentRequestEntity;
-    }
-
-    private PaymentRequestEntity persistPaymentRequestForCardBrand(String cardBrand) {
-        PaymentRequestEntity paymentRequestEntity = aValidPaymentRequestEntityWithRefund()
-                .withGatewayAccountEntity(gatewayAccount)
-                .build();
-       
-        paymentRequestDao.persist(paymentRequestEntity);
-
-        return paymentRequestEntity;
-    }
-
-    private void assertTransactionByExternalId(String expectedExternalId, List<Transaction> searchResult) {
-        assertThat(searchResult.size(), is(1));
-        final String actualExternalId = searchResult.get(0).getExternalId();
-        assertThat(actualExternalId, is(expectedExternalId));
-    }
-
-    private void assertTransactionById(Long expectedId, List<Transaction> searchResult) {
-        assertThat(searchResult.size(), is(1));
-        assertThat(searchResult.get(0).getChargeId(), is(expectedId));
-    }
-
-    private void insertTestAccount() {
-        this.defaultTestAccount = DatabaseFixtures
-                .withDatabaseTestHelper(databaseTestHelper)
-                .aTestAccount()
+    public void searchChargesByReferenceAndEmail_with_under_score() {
+        // since '_' have special meaning in like queries of postgres this was resulting in undesired results
+        withDatabaseTestHelper(databaseTestHelper)
+                .aTestCharge()
+                .withTestAccount(defaultTestAccount)
+                .withReference(ServicePaymentReference.of("under_score_ref"))
+                .withEmail("under_score@mail.com")
                 .insert();
+
+        withDatabaseTestHelper(databaseTestHelper)
+                .aTestCharge()
+                .withTestAccount(defaultTestAccount)
+                .withReference(ServicePaymentReference.of("understand"))
+                .withEmail("undertaker@mail.com")
+                .insert();
+
+        ChargeSearchParams params = new ChargeSearchParams()
+                .withReferenceLike(ServicePaymentReference.of("under_"))
+                .withEmailLike("under_");
+
+        // when
+        List<Transaction> transactions = transactionDao.findAllBy(defaultTestAccount.getAccountId(), params);
+
+        // then
+        assertThat(transactions.size(), is(1));
+
+        Transaction transaction = transactions.get(0);
+        assertThat(transaction.getReference(), is(ServicePaymentReference.of("under_score_ref")));
+        assertThat(transaction.getEmail(), is("under_score@mail.com"));
+    }
+
+    @Test
+    public void searchChargesByReferenceWithPercentSign()  {
+        // since '%' have special meaning in like queries of postgres this was resulting in undesired results
+        withDatabaseTestHelper(databaseTestHelper)
+                .aTestCharge()
+                .withTestAccount(defaultTestAccount)
+                .withReference(ServicePaymentReference.of("percent%ref"))
+                .insert();
+
+        withDatabaseTestHelper(databaseTestHelper)
+                .aTestCharge()
+                .withTestAccount(defaultTestAccount)
+                .withReference(ServicePaymentReference.of("percentref"))
+                .insert();
+
+        ChargeSearchParams params = new ChargeSearchParams()
+                .withReferenceLike(ServicePaymentReference.of("percent%"));
+
+        // when
+        List<Transaction> transactions = transactionDao.findAllBy(defaultTestAccount.getAccountId(), params);
+
+        // then
+        assertThat(transactions.size(), is(1));
+
+        Transaction transaction = transactions.get(0);
+        assertThat(transaction.getReference(), is(ServicePaymentReference.of("percent%ref")));
+    }
+
+    @Test
+    public void searchTransactionByReferenceAndEmailShouldBeCaseInsensitive()  {
+        // fix that the reference and email searches should be case insensitive
+        withDatabaseTestHelper(databaseTestHelper)
+                .aTestCharge()
+                .withTestAccount(defaultTestAccount)
+                .withEmail("email-id@mail.com")
+                .withReference(ServicePaymentReference.of("case-Insensitive-ref"))
+                .insert();
+
+        withDatabaseTestHelper(databaseTestHelper)
+                .aTestCharge()
+                .withTestAccount(defaultTestAccount)
+                .withEmail("EMAIL-ID@MAIL.COM")
+                .withReference(ServicePaymentReference.of("Case-inSENSITIVE-Ref"))
+                .insert();
+
+        ChargeSearchParams params = new ChargeSearchParams()
+                .withReferenceLike(ServicePaymentReference.of("cASe-insEnsiTIve"))
+                .withEmailLike("EMAIL-ID@mail.com");
+
+        // when
+        List<Transaction> transactions = transactionDao.findAllBy(defaultTestAccount.getAccountId(), params);
+
+        // then
+        assertThat(transactions.size(), is(2));
+
+        Transaction transaction = transactions.get(0);
+        assertThat(transaction.getReference(), is(ServicePaymentReference.of("Case-inSENSITIVE-Ref")));
+        assertThat(transaction.getEmail(), is("EMAIL-ID@MAIL.COM"));
+
+        transaction = transactions.get(1);
+        assertThat(transaction.getReference(), is(ServicePaymentReference.of("case-Insensitive-ref")));
+        assertThat(transaction.getEmail(), is("email-id@mail.com"));
+    }
+
+    @Test
+    public void aBasicTestAgainstSqlInjection()  {
+        // given
+        withDatabaseTestHelper(databaseTestHelper)
+                .aTestCharge()
+                .withTestAccount(defaultTestAccount)
+                .withReference(ServicePaymentReference.of("Test reference"))
+                .insert();
+        ChargeSearchParams params = new ChargeSearchParams()
+                .withReferenceLike(ServicePaymentReference.of("reference"));
+        // when passed in a simple reference string
+        List<Transaction> transactions = transactionDao.findAllBy(defaultTestAccount.getAccountId(), params);
+        // then it fetches a single result
+        assertThat(transactions.size(), is(1));
+
+        // when passed in a non existent reference with an sql injected string
+        String sqlInjectionReferenceString = "reffff%' or 1=1 or c.reference like '%1";
+        params = new ChargeSearchParams()
+                .withReferenceLike(ServicePaymentReference.of(sqlInjectionReferenceString));
+        transactions = transactionDao.findAllBy(defaultTestAccount.getAccountId(), params);
+        // then it fetches no result
+        // with a typical sql injection vulnerable query doing this should fetch all results
+        assertThat(transactions.size(), is(0));
+    }
+
+
+    @Test
+    public void searchTransactionByReferenceAndLegacyStatusOnly()  {
+        // given
+        DatabaseFixtures.TestCharge testCharge = insertNewChargeWithId(1L, now());
+        DatabaseFixtures.TestCardDetails testCardDetails = updateCardDetailsForCharge(testCharge);
+        DatabaseFixtures.TestRefund testRefund = insertNewRefundForCharge(testCharge, REFUND_USER_EXTERNAL_ID,2L, now().plusSeconds(1));
+
+        ChargeSearchParams params = new ChargeSearchParams()
+                .withReferenceLike(testCharge.getReference())
+                .addExternalChargeStates(singletonList(EXTERNAL_CREATED.getStatus()));
+
+        // when
+        List<Transaction> transactions = transactionDao.findAllBy(defaultTestAccount.getAccountId(), params);
+
+        // then
+        assertThat(transactions.size(), is(2));
+        Transaction refund = transactions.get(0);
+        Transaction charge = transactions.get(1);
+
+        assertThat(charge.getTransactionType(), is("charge"));
+        assertThat(charge.getChargeId(), is(testCharge.getChargeId()));
+        assertThat(charge.getAmount(), is(testCharge.getAmount()));
+        assertThat(charge.getReference(), is(testCharge.getReference()));
+        assertThat(charge.getDescription(), is(testCharge.getDescription()));
+        assertThat(charge.getStatus(), is(testCharge.getChargeStatus().toString()));
+        assertThat(charge.getCardBrand(), is(testCardDetails.getCardBrand()));
+
+        assertDateMatch(charge.getCreatedDate().toString());
+
+        assertThat(refund.getTransactionType(), is("refund"));
+        assertThat(refund.getChargeId(), is(testCharge.getChargeId()));
+        assertThat(refund.getAmount(), is(testRefund.getAmount()));
+        assertThat(refund.getReference(), is(testCharge.getReference()));
+        assertThat(refund.getDescription(), is(testCharge.getDescription()));
+        assertThat(refund.getStatus(), is(testRefund.getStatus().toString()));
+        assertThat(refund.getCardBrand(), is(testCardDetails.getCardBrand()));
+
+        assertDateMatch(refund.getCreatedDate().toString());
+    }
+
+    @Test
+    public void searchTransactionByReferenceAndStatusAndFromDateAndToDate()  {
+
+        // given
+        DatabaseFixtures.TestCharge testCharge = insertNewChargeWithId(1L, now());
+        DatabaseFixtures.TestCardDetails testCardDetails = updateCardDetailsForCharge(testCharge);
+        DatabaseFixtures.TestRefund testRefund = insertNewRefundForCharge(testCharge, REFUND_USER_EXTERNAL_ID,2L, now().plusSeconds(1));
+
+        ChargeSearchParams params = new ChargeSearchParams()
+                .withReferenceLike(testCharge.getReference())
+                .addExternalChargeStates(singletonList(EXTERNAL_CREATED.getStatus()))
+                .withFromDate(ZonedDateTime.parse(FROM_DATE))
+                .withToDate(ZonedDateTime.parse(TO_DATE));
+
+        // when
+        List<Transaction> transactions = transactionDao.findAllBy(defaultTestAccount.getAccountId(), params);
+
+        // then
+        assertThat(transactions.size(), is(2));
+
+        Transaction transactionRefund = transactions.get(0);
+        assertThat(transactionRefund.getTransactionType(), is("refund"));
+        assertThat(transactionRefund.getChargeId(), is(testCharge.getChargeId()));
+        assertThat(transactionRefund.getAmount(), is(testRefund.getAmount()));
+        assertThat(transactionRefund.getReference(), is(testCharge.getReference()));
+        assertThat(transactionRefund.getDescription(), is(testCharge.getDescription()));
+        assertThat(transactionRefund.getStatus(), is(testRefund.getStatus().toString()));
+        assertThat(transactionRefund.getCardBrand(), is(testCardDetails.getCardBrand()));
+        assertDateMatch(transactionRefund.getCreatedDate().toString());
+
+        Transaction transactionCharge = transactions.get(1);
+        assertThat(transactionCharge.getTransactionType(), is("charge"));
+        assertThat(transactionCharge.getChargeId(), is(testCharge.getChargeId()));
+        assertThat(transactionCharge.getAmount(), is(testCharge.getAmount()));
+        assertThat(transactionCharge.getReference(), is(testCharge.getReference()));
+        assertThat(transactionCharge.getDescription(), is(testCharge.getDescription()));
+        assertThat(transactionCharge.getStatus(), is(testCharge.getChargeStatus().toString()));
+        assertThat(transactionCharge.getCardBrand(), is(testCardDetails.getCardBrand()));
+        assertDateMatch(transactionCharge.getCreatedDate().toString());
+    }
+
+    @Test
+    public void searchChargeByReferenceAndStatusAndFromDate()  {
+
+        // given
+        DatabaseFixtures.TestCharge testCharge = insertNewChargeWithId(1L, now());
+        DatabaseFixtures.TestCardDetails testCardDetails = updateCardDetailsForCharge(testCharge);
+        insertNewRefundForCharge(testCharge, REFUND_USER_EXTERNAL_ID,2L, now().plusSeconds(2));
+
+        ChargeSearchParams params = new ChargeSearchParams()
+                .withTransactionType(PAYMENT)
+                .withReferenceLike(testCharge.getReference())
+                .addExternalChargeStates(singletonList(EXTERNAL_CREATED.getStatus()))
+                .withFromDate(ZonedDateTime.parse(FROM_DATE));
+
+        // when
+        List<Transaction> transactions = transactionDao.findAllBy(defaultTestAccount.getAccountId(), params);
+
+        // then
+        assertThat(transactions.size(), is(1));
+
+        Transaction transactionCharge = transactions.get(0);
+        assertThat(transactionCharge.getTransactionType(), is("charge"));
+        assertThat(transactionCharge.getChargeId(), is(testCharge.getChargeId()));
+        assertThat(transactionCharge.getAmount(), is(testCharge.getAmount()));
+        assertThat(transactionCharge.getReference(), is(testCharge.getReference()));
+        assertThat(transactionCharge.getDescription(), is(testCharge.getDescription()));
+        assertThat(transactionCharge.getStatus(), is(testCharge.getChargeStatus().toString()));
+        assertThat(transactionCharge.getCardBrand(), is(testCardDetails.getCardBrand()));
+        assertDateMatch(transactionCharge.getCreatedDate().toString());
+    }
+
+    @Test
+    public void searchRefundByReferenceAndRefundStatusAndFromDate()  {
+
+        // given
+        DatabaseFixtures.TestCharge testCharge = insertNewChargeWithId(1L, now());
+        DatabaseFixtures.TestCardDetails testCardDetails = updateCardDetailsForCharge(testCharge);
+        DatabaseFixtures.TestRefund testRefund = insertNewRefundForCharge(testCharge, REFUND_USER_EXTERNAL_ID,2L, now().plusSeconds(2));
+
+        ChargeSearchParams params = new ChargeSearchParams()
+                .withTransactionType(REFUND)
+                .withReferenceLike(testCharge.getReference())
+                .addExternalRefundStates(singletonList(EXTERNAL_SUBMITTED.getStatus()))
+                .withFromDate(ZonedDateTime.parse(FROM_DATE));
+
+        // when
+        List<Transaction> transactions = transactionDao.findAllBy(defaultTestAccount.getAccountId(), params);
+
+        // then
+        assertThat(transactions.size(), is(1));
+        Transaction refundTransaction = transactions.get(0);
+
+        assertThat(refundTransaction.getTransactionType(), is("refund"));
+        assertThat(refundTransaction.getChargeId(), is(testCharge.getChargeId()));
+        assertThat(refundTransaction.getAmount(), is(testRefund.getAmount()));
+        assertThat(refundTransaction.getReference(), is(testCharge.getReference()));
+        assertThat(refundTransaction.getDescription(), is(testCharge.getDescription()));
+        assertThat(refundTransaction.getStatus(), is(testRefund.getStatus().toString()));
+        assertThat(refundTransaction.getCardBrand(), is(testCardDetails.getCardBrand()));
+        assertDateMatch(refundTransaction.getCreatedDate().toString());
+    }
+
+    @Test
+    public void searchBySingleChargeStatus() {
+
+        // given
+        DatabaseFixtures.TestCharge charge1 =
+                withDatabaseTestHelper(databaseTestHelper)
+                        .aTestCharge()
+                        .withTestAccount(defaultTestAccount)
+                        .withChargeStatus(ChargeStatus.CREATED)
+                        .withCreatedDate(now())
+                        .insert();
+
+        DatabaseFixtures.TestCharge charge2 =
+                withDatabaseTestHelper(databaseTestHelper)
+                        .aTestCharge()
+                        .withTestAccount(defaultTestAccount)
+                        .withChargeStatus(ENTERING_CARD_DETAILS)
+                        .withCreatedDate(now().plusMinutes(1))
+                        .insert();
+
+
+        withDatabaseTestHelper(databaseTestHelper)
+                .aTestCharge()
+                .withTestAccount(defaultTestAccount)
+                .withChargeStatus(AUTHORISATION_READY)
+                .withCreatedDate(now().plusMinutes(2))
+                .insert();
+
+        withDatabaseTestHelper(databaseTestHelper)
+                .aTestRefund()
+                .withRefundStatus(CREATED)
+                .withTestCharge(charge1)
+                .withCreatedDate(now().plusMinutes(3))
+                .insert();
+
+        withDatabaseTestHelper(databaseTestHelper)
+                .aTestRefund()
+                .withRefundStatus(REFUNDED)
+                .withTestCharge(charge2)
+                .withCreatedDate(now().plusMinutes(4))
+                .insert();
+
+        ChargeSearchParams params = new ChargeSearchParams()
+                .addExternalChargeStates(singletonList(EXTERNAL_STARTED.getStatus()));
+
+        // when
+        List<Transaction> transactions = transactionDao.findAllBy(defaultTestAccount.getAccountId(), params);
+
+        // then
+        assertThat(transactions.size(), is(4));
+        assertThat(transactions.get(0).getTransactionType(), is("refund"));
+        assertThat(transactions.get(0).getStatus(), is(RefundStatus.REFUNDED.getValue()));
+        assertThat(transactions.get(1).getTransactionType(), is("refund"));
+        assertThat(transactions.get(1).getStatus(), is(RefundStatus.CREATED.getValue()));
+        assertThat(transactions.get(2).getTransactionType(), is("charge"));
+        assertThat(transactions.get(2).getStatus(), is(AUTHORISATION_READY.getValue()));
+        assertThat(transactions.get(3).getTransactionType(), is("charge"));
+        assertThat(transactions.get(3).getStatus(), is(ENTERING_CARD_DETAILS.getValue()));
+    }
+
+    @Test
+    public void searchByMultipleChargeStatuses() {
+
+        // given
+        DatabaseFixtures.TestCharge charge1 =
+                withDatabaseTestHelper(databaseTestHelper)
+                        .aTestCharge()
+                        .withTestAccount(defaultTestAccount)
+                        .withChargeStatus(ChargeStatus.CREATED)
+                        .withCreatedDate(now())
+                        .insert();
+
+        DatabaseFixtures.TestCharge charge2 =
+                withDatabaseTestHelper(databaseTestHelper)
+                        .aTestCharge()
+                        .withTestAccount(defaultTestAccount)
+                        .withChargeStatus(ENTERING_CARD_DETAILS)
+                        .withCreatedDate(now().plusMinutes(1))
+                        .insert();
+
+
+        withDatabaseTestHelper(databaseTestHelper)
+                .aTestCharge()
+                .withTestAccount(defaultTestAccount)
+                .withChargeStatus(AUTHORISATION_READY)
+                .withCreatedDate(now().plusMinutes(2))
+                .insert();
+
+        withDatabaseTestHelper(databaseTestHelper)
+                .aTestRefund()
+                .withRefundStatus(CREATED)
+                .withTestCharge(charge1)
+                .withCreatedDate(now().plusMinutes(3))
+                .insert();
+
+        withDatabaseTestHelper(databaseTestHelper)
+                .aTestRefund()
+                .withRefundStatus(REFUNDED)
+                .withTestCharge(charge2)
+                .withCreatedDate(now().plusMinutes(4))
+                .insert();
+
+        ChargeSearchParams params = new ChargeSearchParams()
+                .addExternalChargeStates(singletonList(EXTERNAL_STARTED.getStatus()))
+                .addExternalChargeStates(singletonList(EXTERNAL_CREATED.getStatus()));
+
+        // when
+        List<Transaction> transactions = transactionDao.findAllBy(defaultTestAccount.getAccountId(), params);
+
+        // then
+        assertThat(transactions.size(), is(5));
+        assertThat(transactions.get(0).getTransactionType(), is("refund"));
+        assertThat(transactions.get(0).getStatus(), is(RefundStatus.REFUNDED.getValue()));
+        assertThat(transactions.get(1).getTransactionType(), is("refund"));
+        assertThat(transactions.get(1).getStatus(), is(RefundStatus.CREATED.getValue()));
+        assertThat(transactions.get(2).getTransactionType(), is("charge"));
+        assertThat(transactions.get(2).getStatus(), is(AUTHORISATION_READY.getValue()));
+        assertThat(transactions.get(3).getTransactionType(), is("charge"));
+        assertThat(transactions.get(3).getStatus(), is(ENTERING_CARD_DETAILS.getValue()));
+        assertThat(transactions.get(4).getTransactionType(), is("charge"));
+        assertThat(transactions.get(4).getStatus(), is(ChargeStatus.CREATED.getValue()));
+    }
+
+    @Test
+    public void searchBySingleChargeStatusAndTransactionType() {
+        // given
+        DatabaseFixtures.TestCharge charge1 =
+                withDatabaseTestHelper(databaseTestHelper)
+                        .aTestCharge()
+                        .withTestAccount(defaultTestAccount)
+                        .withChargeStatus(ChargeStatus.CREATED)
+                        .withCreatedDate(now())
+                        .insert();
+
+        DatabaseFixtures.TestCharge charge2 =
+                withDatabaseTestHelper(databaseTestHelper)
+                        .aTestCharge()
+                        .withTestAccount(defaultTestAccount)
+                        .withChargeStatus(ENTERING_CARD_DETAILS)
+                        .withCreatedDate(now().plusMinutes(1))
+                        .insert();
+
+
+        withDatabaseTestHelper(databaseTestHelper)
+                .aTestCharge()
+                .withTestAccount(defaultTestAccount)
+                .withChargeStatus(AUTHORISATION_READY)
+                .withCreatedDate(now().plusMinutes(2))
+                .insert();
+
+        withDatabaseTestHelper(databaseTestHelper)
+                .aTestRefund()
+                .withRefundStatus(CREATED)
+                .withTestCharge(charge1)
+                .withCreatedDate(now().plusMinutes(3))
+                .insert();
+
+        withDatabaseTestHelper(databaseTestHelper)
+                .aTestRefund()
+                .withRefundStatus(REFUNDED)
+                .withTestCharge(charge2)
+                .withCreatedDate(now().plusMinutes(4))
+                .insert();
+
+        ChargeSearchParams params = new ChargeSearchParams()
+                .withTransactionType(PAYMENT)
+                .addExternalChargeStates(singletonList(EXTERNAL_STARTED.getStatus()));
+
+        // when
+        List<Transaction> transactions = transactionDao.findAllBy(defaultTestAccount.getAccountId(), params);
+
+        // then
+        assertThat(transactions.size(), is(2));
+        assertThat(transactions.get(0).getTransactionType(), is("charge"));
+        assertThat(transactions.get(0).getStatus(), is(AUTHORISATION_READY.getValue()));
+        assertThat(transactions.get(1).getTransactionType(), is("charge"));
+        assertThat(transactions.get(1).getStatus(), is(ENTERING_CARD_DETAILS.getValue()));
+    }
+
+    @Test
+    public void searchBySingleRefundStatus() {
+
+        // given
+        DatabaseFixtures.TestCharge charge1 =
+                withDatabaseTestHelper(databaseTestHelper)
+                        .aTestCharge()
+                        .withTestAccount(defaultTestAccount)
+                        .withChargeStatus(ChargeStatus.CREATED)
+                        .withCreatedDate(now())
+                        .insert();
+
+        DatabaseFixtures.TestCharge charge2 =
+                withDatabaseTestHelper(databaseTestHelper)
+                        .aTestCharge()
+                        .withTestAccount(defaultTestAccount)
+                        .withChargeStatus(ENTERING_CARD_DETAILS)
+                        .withCreatedDate(now().plusMinutes(1))
+                        .insert();
+
+
+        withDatabaseTestHelper(databaseTestHelper)
+                .aTestCharge()
+                .withTestAccount(defaultTestAccount)
+                .withChargeStatus(AUTHORISATION_READY)
+                .withCreatedDate(now().plusMinutes(2))
+                .insert();
+
+        withDatabaseTestHelper(databaseTestHelper)
+                .aTestRefund()
+                .withRefundStatus(CREATED)
+                .withTestCharge(charge1)
+                .withCreatedDate(now().plusMinutes(3))
+                .insert();
+
+        withDatabaseTestHelper(databaseTestHelper)
+                .aTestRefund()
+                .withRefundStatus(REFUNDED)
+                .withTestCharge(charge2)
+                .withCreatedDate(now().plusMinutes(4))
+                .insert();
+
+        ChargeSearchParams params = new ChargeSearchParams()
+                .addExternalRefundStates(singletonList(EXTERNAL_SUCCESS.getStatus()));
+
+        // when
+        List<Transaction> transactions = transactionDao.findAllBy(defaultTestAccount.getAccountId(), params);
+
+        // then
+        assertThat(transactions.size(), is(4));
+        assertThat(transactions.get(0).getTransactionType(), is("refund"));
+        assertThat(transactions.get(0).getStatus(), is(RefundStatus.REFUNDED.getValue()));
+        assertThat(transactions.get(1).getTransactionType(), is("charge"));
+        assertThat(transactions.get(1).getStatus(), is(AUTHORISATION_READY.getValue()));
+        assertThat(transactions.get(2).getTransactionType(), is("charge"));
+        assertThat(transactions.get(2).getStatus(), is(ENTERING_CARD_DETAILS.getValue()));
+        assertThat(transactions.get(3).getTransactionType(), is("charge"));
+        assertThat(transactions.get(3).getStatus(), is(ChargeStatus.CREATED.getValue()));
+    }
+
+    @Test
+    public void searchByMultipleRefundStatus() {
+
+        // given
+        DatabaseFixtures.TestCharge charge1 =
+                withDatabaseTestHelper(databaseTestHelper)
+                        .aTestCharge()
+                        .withTestAccount(defaultTestAccount)
+                        .withChargeStatus(ChargeStatus.CREATED)
+                        .withCreatedDate(now())
+                        .insert();
+
+        DatabaseFixtures.TestCharge charge2 =
+                withDatabaseTestHelper(databaseTestHelper)
+                        .aTestCharge()
+                        .withTestAccount(defaultTestAccount)
+                        .withChargeStatus(ENTERING_CARD_DETAILS)
+                        .withCreatedDate(now().plusMinutes(1))
+                        .insert();
+
+
+        withDatabaseTestHelper(databaseTestHelper)
+                .aTestCharge()
+                .withTestAccount(defaultTestAccount)
+                .withChargeStatus(AUTHORISATION_READY)
+                .withCreatedDate(now().plusMinutes(2))
+                .insert();
+
+        withDatabaseTestHelper(databaseTestHelper)
+                .aTestRefund()
+                .withRefundStatus(CREATED)
+                .withTestCharge(charge1)
+                .withCreatedDate(now().plusMinutes(3))
+                .insert();
+
+        withDatabaseTestHelper(databaseTestHelper)
+                .aTestRefund()
+                .withRefundStatus(REFUNDED)
+                .withTestCharge(charge2)
+                .withCreatedDate(now().plusMinutes(4))
+                .insert();
+
+        ChargeSearchParams params = new ChargeSearchParams()
+                .addExternalRefundStates(singletonList(EXTERNAL_SUBMITTED.getStatus()))
+                .addExternalRefundStates(singletonList(EXTERNAL_SUCCESS.getStatus()));
+
+        // when
+        List<Transaction> transactions = transactionDao.findAllBy(defaultTestAccount.getAccountId(), params);
+
+        // then
+        assertThat(transactions.size(), is(5));
+        assertThat(transactions.get(0).getTransactionType(), is("refund"));
+        assertThat(transactions.get(0).getStatus(), is(RefundStatus.REFUNDED.getValue()));
+        assertThat(transactions.get(1).getTransactionType(), is("refund"));
+        assertThat(transactions.get(1).getStatus(), is(RefundStatus.CREATED.getValue()));
+        assertThat(transactions.get(2).getTransactionType(), is("charge"));
+        assertThat(transactions.get(2).getStatus(), is(AUTHORISATION_READY.getValue()));
+        assertThat(transactions.get(3).getTransactionType(), is("charge"));
+        assertThat(transactions.get(3).getStatus(), is(ENTERING_CARD_DETAILS.getValue()));
+        assertThat(transactions.get(4).getTransactionType(), is("charge"));
+        assertThat(transactions.get(4).getStatus(), is(ChargeStatus.CREATED.getValue()));
+    }
+
+    @Test
+    public void searchBySingleRefundStatusAndTransactionType() {
+
+        // given
+        DatabaseFixtures.TestCharge charge1 =
+                withDatabaseTestHelper(databaseTestHelper)
+                        .aTestCharge()
+                        .withTestAccount(defaultTestAccount)
+                        .withChargeStatus(ChargeStatus.CREATED)
+                        .withCreatedDate(now())
+                        .insert();
+
+        DatabaseFixtures.TestCharge charge2 =
+                withDatabaseTestHelper(databaseTestHelper)
+                        .aTestCharge()
+                        .withTestAccount(defaultTestAccount)
+                        .withChargeStatus(ENTERING_CARD_DETAILS)
+                        .withCreatedDate(now().plusMinutes(1))
+                        .insert();
+
+
+        withDatabaseTestHelper(databaseTestHelper)
+                .aTestCharge()
+                .withTestAccount(defaultTestAccount)
+                .withChargeStatus(AUTHORISATION_READY)
+                .withCreatedDate(now().plusMinutes(2))
+                .insert();
+
+        withDatabaseTestHelper(databaseTestHelper)
+                .aTestRefund()
+                .withRefundStatus(CREATED)
+                .withTestCharge(charge1)
+                .withCreatedDate(now().plusMinutes(3))
+                .insert();
+
+        withDatabaseTestHelper(databaseTestHelper)
+                .aTestRefund()
+                .withRefundStatus(REFUNDED)
+                .withTestCharge(charge2)
+                .withCreatedDate(now().plusMinutes(4))
+                .insert();
+        ChargeSearchParams params = new ChargeSearchParams()
+                .withTransactionType(REFUND)
+                .addExternalRefundStates(singletonList(EXTERNAL_SUCCESS.getStatus()));
+
+        // when
+        List<Transaction> transactions = transactionDao.findAllBy(defaultTestAccount.getAccountId(), params);
+
+        // then
+        assertThat(transactions.size(), is(1));
+        assertThat(transactions.get(0).getTransactionType(), is("refund"));
+        assertThat(transactions.get(0).getStatus(), is(RefundStatus.REFUNDED.getValue()));
+    }
+
+    @Test
+    public void searchChargeByReferenceAndStatusAndEmailAndCardBrandAndFromDateAndToDate()  {
+        // given
+        DatabaseFixtures.TestCharge testCharge = insertNewChargeWithId(1L, now());
+        DatabaseFixtures.TestCardDetails testCardDetails = updateCardDetailsForCharge(testCharge);
+        insertNewRefundForCharge(testCharge, REFUND_USER_EXTERNAL_ID, 2L, now().plusSeconds(2));
+
+        ChargeSearchParams params = new ChargeSearchParams()
+                .withTransactionType(PAYMENT)
+                .withReferenceLike(testCharge.getReference())
+                .addExternalChargeStates(singletonList(EXTERNAL_CREATED.getStatus()))
+                .withCardBrand(testCardDetails.getCardBrand())
+                .withEmailLike(testCharge.getEmail())
+                .withFromDate(ZonedDateTime.parse(FROM_DATE))
+                .withToDate(ZonedDateTime.parse(TO_DATE));
+
+        // when
+        List<Transaction> transactions = transactionDao.findAllBy(defaultTestAccount.getAccountId(), params);
+
+        // then
+        assertThat(transactions.size(), is(1));
+        Transaction charge = transactions.get(0);
+
+        assertThat(charge.getTransactionType(), is("charge"));
+        assertThat(charge.getChargeId(), is(testCharge.getChargeId()));
+        assertThat(charge.getAmount(), is(testCharge.getAmount()));
+        assertThat(charge.getReference(), is(testCharge.getReference()));
+        assertThat(charge.getDescription(), is(testCharge.getDescription()));
+        assertThat(charge.getStatus(), is(testCharge.getChargeStatus().toString()));
+        assertThat(charge.getCardBrand(), is(testCardDetails.getCardBrand()));
+        assertThat(charge.getUserExternalId(), is(nullValue()));
+        assertDateMatch(charge.getCreatedDate().toString());
+    }
+
+    @Test
+    public void searchChargeByReferenceAndStatusAndToDate()  {
+
+        // given
+        DatabaseFixtures.TestCharge testCharge = insertNewChargeWithId(1L, now());
+        DatabaseFixtures.TestCardDetails testCardDetails = updateCardDetailsForCharge(testCharge);
+        insertNewRefundForCharge(testCharge, REFUND_USER_EXTERNAL_ID, 2L, now().plusSeconds(2));
+
+        ChargeSearchParams params = new ChargeSearchParams()
+                .withTransactionType(PAYMENT)
+                .withReferenceLike(testCharge.getReference())
+                .addExternalChargeStates(singletonList(EXTERNAL_CREATED.getStatus()))
+                .withToDate(ZonedDateTime.parse(TO_DATE));
+
+        // when
+        List<Transaction> transactions = transactionDao.findAllBy(defaultTestAccount.getAccountId(), params);
+
+        // then
+        assertThat(transactions.size(), is(1));
+        Transaction charge = transactions.get(0);
+
+        assertThat(charge.getTransactionType(), is("charge"));
+        assertThat(charge.getChargeId(), is(testCharge.getChargeId()));
+        assertThat(charge.getAmount(), is(testCharge.getAmount()));
+        assertThat(charge.getReference(), is(testCharge.getReference()));
+        assertThat(charge.getDescription(), is(DEFAULT_TEST_CHARGE_DESCRIPTION));
+        assertThat(charge.getStatus(), is(testCharge.getChargeStatus().toString()));
+        assertThat(charge.getCardBrand(), is(testCardDetails.getCardBrand()));
+        assertThat(charge.getUserExternalId(), is(nullValue()));
+
+        assertDateMatch(charge.getCreatedDate().toString());
+    }
+
+    @Test
+    public void searchTransactionByReferenceAndStatusAndFromDate_ShouldReturnZeroIfDateIsNotInRange()  {
+
+        // given
+        DatabaseFixtures.TestCharge testCharge = insertNewChargeWithId(1L, now().plusHours(1));
+        insertNewRefundForCharge(testCharge, REFUND_USER_EXTERNAL_ID, 2L, now().plusHours(2));
+
+        ChargeSearchParams params = new ChargeSearchParams()
+                .withReferenceLike(testCharge.getReference())
+                .addExternalChargeStates(singletonList(EXTERNAL_CREATED.getStatus()))
+                .withFromDate(ZonedDateTime.parse(TO_DATE));
+
+        // when
+        List<Transaction> transactions = transactionDao.findAllBy(defaultTestAccount.getAccountId(), params);
+
+        // then
+        assertThat(transactions.size(), is(0));
+    }
+
+    @Test
+    public void searchTransactionByReferenceAndStatusAndToDate_ShouldReturnZeroIfToDateIsNotInRange()  {
+
+        // given
+        DatabaseFixtures.TestCharge testCharge = insertNewChargeWithId(1L, now().plusHours(1));
+        insertNewRefundForCharge(testCharge, REFUND_USER_EXTERNAL_ID, 2L, now().plusHours(2));
+
+        ChargeSearchParams params = new ChargeSearchParams()
+                .withReferenceLike(testCharge.getReference())
+                .addExternalChargeStates(singletonList(EXTERNAL_CREATED.getStatus()))
+                .withToDate(ZonedDateTime.parse(FROM_DATE));
+
+        // when
+        List<Transaction> transactions = transactionDao.findAllBy(defaultTestAccount.getAccountId(), params);
+
+        // then
+        assertThat(transactions.size(), is(0));
+    }
+
+    private void assertDateMatch(String createdDateString) {
+        assertDateMatch(DateTimeUtils.toUTCZonedDateTime(createdDateString).get());
+    }
+
+    private void assertDateMatch(ZonedDateTime createdDateTime) {
+        assertThat(createdDateTime, within(1, ChronoUnit.MINUTES, ZonedDateTime.now()));
+    }
+
+    private DatabaseFixtures.TestCharge insertNewChargeWithId(long amount, ZonedDateTime creationDate) {
+        return
+                withDatabaseTestHelper(databaseTestHelper)
+                        .aTestCharge()
+                        .withDescription(DEFAULT_TEST_CHARGE_DESCRIPTION)
+                        .withReference(ServicePaymentReference.of(DEFAULT_TEST_CHARGE_REFERENCE))
+                        .withAmount(amount)
+                        .withTestAccount(defaultTestAccount)
+                        .withCreatedDate(creationDate)
+                        .insert();
+    }
+
+    private DatabaseFixtures.TestCardDetails updateCardDetailsForCharge(DatabaseFixtures.TestCharge charge) {
+        return updateCardDetailsForCharge(charge, "visa");
+    }
+
+    private DatabaseFixtures.TestCardDetails updateCardDetailsForCharge(DatabaseFixtures.TestCharge charge, String cardBrand) {
+        return new DatabaseFixtures(databaseTestHelper)
+                .validTestCardDetails()
+                .withChargeId(charge.chargeId)
+                .withCardBrand(cardBrand)
+                .update();
+    }
+
+    private DatabaseFixtures.TestRefund insertNewRefundForCharge(DatabaseFixtures.TestCharge charge, String userExternalId, long amount, ZonedDateTime creationDate) {
+        return
+                withDatabaseTestHelper(databaseTestHelper)
+                        .aTestRefund()
+                        .withAmount(amount)
+                        .withTestCharge(charge)
+                        .withCreatedDate(creationDate)
+                        .withSubmittedBy(userExternalId)
+                        .insert();
     }
 }
