+++ src/test/java/uk/gov/pay/connector/model/domain/PaymentRequestEntityFixture.java
@@ -1,7 +1,6 @@
 package uk.gov.pay.connector.model.domain;
 
 import uk.gov.pay.connector.model.ServicePaymentReference;
-import uk.gov.pay.connector.model.domain.transaction.TransactionEntity;
 import uk.gov.pay.connector.util.RandomIdGenerator;
 
 import java.time.ZoneId;
@@ -11,8 +10,6 @@ import java.util.List;
 
 import static java.util.Arrays.asList;
 import static uk.gov.pay.connector.model.domain.ChargeEntityFixture.defaultGatewayAccountEntity;
-import static uk.gov.pay.connector.model.domain.transaction.ChargeTransactionEntityBuilder.aChargeTransactionEntity;
-import static uk.gov.pay.connector.model.domain.transaction.RefundTransactionEntityBuilder.aRefundTransactionEntity;
 
 public class PaymentRequestEntityFixture {
 
@@ -23,7 +20,6 @@ public class PaymentRequestEntityFixture {
     private ServicePaymentReference reference = ServicePaymentReference.of(RandomIdGenerator.newId());
     private GatewayAccountEntity gatewayAccountEntity = defaultGatewayAccountEntity();
     private ZonedDateTime createdDate = ZonedDateTime.now(ZoneId.of("UTC"));
-    private List<TransactionEntity> transactions = asList(aChargeTransactionEntity().build());
 
     public static PaymentRequestEntityFixture aValidPaymentRequestEntity() {
         return new PaymentRequestEntityFixture();
@@ -31,7 +27,7 @@ public class PaymentRequestEntityFixture {
 
     public static PaymentRequestEntityFixture aValidPaymentRequestEntityWithRefund() {
         return PaymentRequestEntityFixture.aValidPaymentRequestEntity()
-                .withTransactions(aChargeTransactionEntity().build(), aRefundTransactionEntity().build());
+                ;
     }
 
     public PaymentRequestEntity build() {
@@ -43,18 +39,12 @@ public class PaymentRequestEntityFixture {
         entity.setDescription(this.description);
         entity.setCreatedDate(this.createdDate);
         entity.setAmount(this.amount);
-        entity.setTransactions(transactions);
 
         return entity;
     }
 
     public PaymentRequestEntityFixture withGatewayAccountEntity(GatewayAccountEntity gatewayAccount) {
         this.gatewayAccountEntity = gatewayAccount;
-        return this;
-    }
-
-    public PaymentRequestEntityFixture withTransactions(TransactionEntity... transactions) {
-        this.transactions = Arrays.asList(transactions);
         return this;
     }
