+++ src/test/java/uk/gov/pay/connector/service/NotificationServiceTest.java
@@ -78,9 +78,6 @@ public class NotificationServiceTest {
     @Mock
     private ChargeStatusUpdater mockedChargeStatusUpdater;
 
-    @Mock
-    private RefundStatusUpdater mockedRefundUpdater;
-
     @Before
     public void setUp() {
         when(mockedPaymentProvider.getPaymentGatewayName()).thenReturn(SANDBOX);
@@ -92,7 +89,7 @@ public class NotificationServiceTest {
 
         when(mockedPaymentProvider.verifyNotification(any(Notification.class), any(GatewayAccountEntity.class))).thenReturn(true);
 
-        notificationService = new NotificationService(mockedChargeDao, mockedChargeEventDao, mockedRefundDao, mockedPaymentProviders, mockDnsUtils, mockedChargeStatusUpdater, mockedRefundUpdater);
+        notificationService = new NotificationService(mockedChargeDao, mockedChargeEventDao, mockedRefundDao, mockedPaymentProviders, mockDnsUtils, mockedChargeStatusUpdater);
     }
 
     private Notifications<Pair<String, Boolean>> createNotificationFor(String transactionId, String reference, Pair<String, Boolean> status) {
@@ -353,7 +350,6 @@ public class NotificationServiceTest {
 
         verify(mockedRefundDao).findByProviderAndReference(SANDBOX.getName(), reference);
         verify(mockedRefundEntity).setStatus(REFUNDED);
-        verify(mockedRefundUpdater).updateRefundTransactionStatus(SANDBOX, reference, RefundStatus.REFUNDED);
         verifyNoMoreInteractions(ignoreStubs(mockedChargeDao));
     }
